.cursor/requirements.txt
pyyaml>=6.0,<7.0
pytest>=8.0,<9.0
